+++ controllers/podplacement/suite_test.go
@@ -40,6 +40,7 @@ import (
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/client-go/rest"
+
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/envtest"
@@ -48,8 +49,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 	"sigs.k8s.io/controller-runtime/pkg/webhook"
 
+	testingutils "github.com/openshift/multiarch-manager-operator/pkg/testing/framework"
 	"github.com/openshift/multiarch-manager-operator/pkg/testing/image/fake/registry"
-	testingutils "github.com/openshift/multiarch-manager-operator/pkg/testing/utils"
 	"github.com/openshift/multiarch-manager-operator/pkg/utils"
 	//+kubebuilder:scaffold:imports
 )
